src/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
gitdb==4.0.12
GitPython==3.1.41
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Markdown==3.8
packaging==25.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
setuptools==75.6.0
smmap==5.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
